refactor(commonjs): extract logic to rewrite imports
