+++ packages/commonjs/src/handle-require-expressions.js
@@ -2,7 +2,13 @@ import { dirname, resolve } from 'path';
 
 import { sync as nodeResolveSync } from 'resolve';
 
-import { DYNAMIC_JSON_PREFIX, DYNAMIC_REGISTER_PREFIX } from './helpers';
+import { isLocallyShadowed } from './ast-utils';
+
+import {
+  DYNAMIC_JSON_PREFIX,
+  DYNAMIC_REGISTER_PREFIX,
+  getVirtualPathForDynamicRequirePath
+} from './helpers';
 import { normalizePathSlashes } from './utils';
 
 export function isRequireStatement(node, scope) {
@@ -89,7 +95,7 @@ function hasDynamicModuleForPath(source, id, dynamicRequireModuleSet) {
   return false;
 }
 
-export function shouldUseSimulatedRequire(required, id, dynamicRequireModuleSet) {
+function shouldUseSimulatedRequire(required, id, dynamicRequireModuleSet) {
   return (
     hasDynamicModuleForPath(required.source, id, dynamicRequireModuleSet) &&
     // We only do `commonjsRequire` for json if it's the `commonjsRegister` call.
@@ -103,6 +109,15 @@ export function getRequireHandlers(id, dynamicRequireModuleSet) {
   const requiredBySource = Object.create(null);
   const requiredByNode = new Map();
   const requireExpressionsWithUsedReturnValue = [];
+
+  function addRequireStatement(node, scope, usesReturnValue) {
+    const required = getRequired(node);
+    requiredByNode.set(node, { scope, required });
+    if (usesReturnValue) {
+      required.nodesUsingRequired.push(node);
+      requireExpressionsWithUsedReturnValue.push(node);
+    }
+  }
 
   function getRequired(node) {
     let sourceId = getRequireStringArg(node);
@@ -136,21 +151,89 @@ export function getRequireHandlers(id, dynamicRequireModuleSet) {
     return requiredBySource[sourceId];
   }
 
-  function addRequireStatement(node, scope, usesReturnValue) {
-    const required = getRequired(node);
-    requiredByNode.set(node, { scope, required });
-    if (usesReturnValue) {
-      required.nodesUsingRequired.push(node);
-      requireExpressionsWithUsedReturnValue.push(node);
+  function rewriteRequireExpressions(
+    magicString,
+    topLevelDeclarations,
+    topLevelRequireDeclarators,
+    reassignedNames,
+    commonDir,
+    uses,
+    HELPERS_NAME
+  ) {
+    // Determine the used names and removed declarators
+    const removedDeclarators = new Set();
+    for (const declarator of topLevelRequireDeclarators) {
+      const { required } = requiredByNode.get(declarator.init);
+      if (!(required.name || required.isDynamic)) {
+        const potentialName = declarator.id.name;
+        if (
+          !reassignedNames.has(potentialName) &&
+          !required.nodesUsingRequired.some((node) =>
+            isLocallyShadowed(potentialName, requiredByNode.get(node).scope)
+          )
+        ) {
+          required.name = potentialName;
+          removedDeclarators.add(declarator);
+        }
+      }
+    }
+
+    // Determine names where they are still missing and rewrite expressions
+    let uid = 0;
+    for (const requireExpression of requireExpressionsWithUsedReturnValue) {
+      const { required } = requiredByNode.get(requireExpression);
+      if (shouldUseSimulatedRequire(required, id, dynamicRequireModuleSet)) {
+        magicString.overwrite(
+          requireExpression.start,
+          requireExpression.end,
+          `${HELPERS_NAME}.commonjsRequire(${JSON.stringify(
+            getVirtualPathForDynamicRequirePath(normalizePathSlashes(required.source), commonDir)
+          )}, ${JSON.stringify(
+            dirname(id) === '.'
+              ? null /* default behavior */
+              : getVirtualPathForDynamicRequirePath(normalizePathSlashes(dirname(id)), commonDir)
+          )})`
+        );
+        // eslint-disable-next-line no-param-reassign
+        uses.commonjsHelpers = true;
+      } else {
+        if (!required.name) {
+          let potentialName;
+          const isUsedName = (node) => requiredByNode.get(node).scope.contains(potentialName);
+          do {
+            potentialName = `require$$${uid}`;
+            uid += 1;
+          } while (required.nodesUsingRequired.some(isUsedName));
+          required.name = potentialName;
+        }
+        magicString.overwrite(requireExpression.start, requireExpression.end, required.name);
+      }
+    }
+
+    // Rewrite declarations, checking which declarators can be removed by their init
+    for (const declaration of topLevelDeclarations) {
+      let keepDeclaration = false;
+      let [{ start }] = declaration.declarations;
+      for (const declarator of declaration.declarations) {
+        if (removedDeclarators.has(declarator)) {
+          magicString.remove(start, declarator.end);
+        } else if (!keepDeclaration) {
+          magicString.remove(start, declarator.start);
+          keepDeclaration = true;
+        }
+        start = declarator.end;
+      }
+      if (!keepDeclaration) {
+        magicString.remove(declaration.start, declaration.end);
+      }
     }
   }
 
   return {
     addRequireStatement,
     dynamicRegisterSources,
-    requiredByNode,
     requiredBySource,
     requiredSources,
-    requireExpressionsWithUsedReturnValue
+    rewriteRequireExpressions
   };
 }
